+++ dlib/memory_manager_global/memory_manager_global_kernel_abstract.h
@@ -81,7 +81,7 @@ namespace dlib
                       get_global_memory_manager().
             !*/
 
-            size_t get_number_of_allocations (
+            unsigned long get_number_of_allocations (
             ) const;
             /*!
                 ensures
@@ -122,7 +122,7 @@ namespace dlib
             !*/
 
             T* allocate_array (
-                size_t size
+                unsigned long size
             );
             /*!
                 ensures
