+++ dlib/memory_manager_stateless/memory_manager_stateless_kernel_2.h
@@ -60,7 +60,7 @@ namespace dlib
             }
 
             T* allocate_array (
-                size_t size
+                unsigned long size
             ) 
             { 
                 auto_mutex M(global_mutex());
